+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## WIP
+
+### Fixed
+
+- Fix support for more than two digits in Countdown component
+
 ## 3.0.0 beta 22 (April 24, 2017)
 
 ### Added